+++ ProjekatWebPortal/Views/Home/UserManager.cshtml
@@ -30,9 +30,9 @@
                 <td>David</td>
                 <td>Bozic</td>
                 <td class="table-btns">
-                    <a asp-controller="Home" asp-action="Info" class="btn btn-sm btn-outline-primary"><i class="bi bi-info-circle-fill"></i> Info</a>
-                    <a asp-controller="Home" asp-action="Izmeni" class="btn btn-sm btn-outline-secondary"><i class="bi bi-pencil"></i> Izmeni</a>
-                    <a asp-controller="Home" asp-action="Obrisi" class="btn btn-sm btn-outline-danger"><i class="bi bi-trash3"></i> Obriši</a>
+                    <a asp-controller="Users" asp-action="Info" class="btn btn-sm btn-outline-primary"><i class="bi bi-info-circle-fill"></i> Info</a>
+                    <a asp-controller="Users" asp-action="Edit" class="btn btn-sm btn-outline-secondary"><i class="bi bi-pencil"></i> Izmeni</a>
+                    <a asp-controller="Users" asp-action="Delete" class="btn btn-sm btn-outline-danger"><i class="bi bi-trash3"></i> Obriši</a>
                 </td>
             </tr>
             <tr>
